requirements.txt
streamlit
google-generativeai
Pillow
requests
gTTS
